Fix: Fix deleting non-existing image file issue and validation messages

+++ utils/file.js
@@ -1,6 +1,7 @@
-const { unlink } = require("fs");
+const { unlink, existsSync } = require("fs");
 
 exports.deleteFile = (filePath) => {
+  if (!existsSync(filePath)) return;
   unlink(filePath, (error) => {
     if (error) throw error;
   });
